Refactor updateGlobalState as a core helper
