+++ packages/core/index.js
@@ -2,7 +2,6 @@ export * from './src/js/Breakpoint';
 export * from './src/js/Collection';
 export * from './src/js/FocusTrap';
 
-export * from './src/js/accessibility';
 export * from './src/js/addClass';
 export * from './src/js/camelCase';
 export * from './src/js/getConfig';
@@ -10,6 +9,9 @@ export * from './src/js/hasClass';
 export * from './src/js/hyphenCase';
 export * from './src/js/localStore';
 export * from './src/js/removeClass';
+export * from './src/js/setInert';
+export * from './src/js/setOverflowHidden';
 export * from './src/js/teleport';
 export * from './src/js/toggleClass';
 export * from './src/js/transition';
+export * from './src/js/updateGlobalState';
