+++ packages/core/src/js/setInert.js
@@ -12,16 +12,3 @@ export const setInert = (state, selector) => {
     });
   }
 };
-
-export const setOverflowHidden = (state, selector) => {
-  if (selector) {
-    const els = document.querySelectorAll(selector);
-    els.forEach((el) => {
-      if (state) {
-        el.style.overflow = 'hidden';
-      } else {
-        el.style.removeProperty('overflow');
-      }
-    });
-  }
-};
