+++ packages/modal/src/js/helpers/index.js
@@ -2,5 +2,4 @@ export * from './getModal';
 export * from './getModalElements';
 export * from './getModalID';
 export * from './updateFocusState';
-export * from './updateGlobalState';
 export * from './updateStackIndex';
